REACT: #19 Fix for React 18 causing a crash when using the usePlexus hook on severside
