+++ packages/plexus-react/package.json
@@ -10,13 +10,14 @@
 	"license": "MIT",
 	"peerDependencies": {
 		"@plexusjs/core": "~1.0.0",
-		"react": "^18.0.0"
+		"react": "^18.2"
 	},
 	"devDependencies": {
 		"@types/react": "^16.9.49",
-		"@types/react-test-renderer": "^16.9.3",
+		"@types/react-test-renderer": "^18.0.0",
+		"@types/use-sync-external-store": "^0.0.3",
 		"react": "^18.2.0",
-		"react-test-renderer": "^16.13.1"
+		"react-test-renderer": "^18.2.0"
 	},
 	"scripts": {
 		"build": "tsc --build --force",
@@ -25,7 +26,6 @@
 	},
 	"gitHead": "48b37fe6945b413ee01261a57e744f71aa39d6fc",
 	"dependencies": {
-		"@types/use-sync-external-store": "^0.0.3",
 		"use-sync-external-store": "^1.2.0"
 	}
 }
